Split test_invalid_request into two modules
